requirements.txt
altair==5.1.2
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.5.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.1.0
blinker==1.6.2
cachetools==5.3.2
cattrs==23.1.2
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.4.1
cmdstanpy==1.2.0
comm==0.2.0
contourpy==1.2.0
cron-descriptor==1.4.0
cryptography==41.0.3
cycler==0.12.1
Cython==3.0.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.3
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-mongodb-engine==0.6.0
django-redis==5.4.0
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djangotoolbox==1.8.0
djongo==1.3.6
dnspython==2.2.1
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.103.1
fastjsonschema==2.18.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.44.0
fqdn==1.5.1
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
holidays==0.36
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
interchange==2021.0.4
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
kombu==5.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
neo4j==5.11.0
neo4j-driver==4.4.10
neobolt==1.7.17
neomodel==5.1.1
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.1
openmeteo_requests==1.1.0
openmeteo_sdk==1.5.0
overrides==7.4.0
packaging==23.1
pandas==2.1.2
pandocfilters==1.5.0
pansi==2020.7.3
parso==0.8.3
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.11.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
prophet==1.1.5
protobuf==4.25.0
psutil==5.9.6
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyclean==2.7.5
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.15.1
PyJWT==2.8.0
# pykerberos==1.2.4
pymongo==3.12.1
pyparsing==3.1.1
pyspnego==0.9.2
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-snappy==0.6.1
pytz==2022.1
pywinrm==0.4.3
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.0
QtPy==2.4.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-cache==1.1.0
requests-ntlm==1.2.0
retry-requests==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.12.0
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.2.4
stack-data==0.6.3
stanio==0.3.0
starlette==0.27.0
streamlit==1.28.1
tenacity==8.2.3
terminado==0.18.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.0.4
uvicorn==0.23.2
validators==0.22.0
vine==5.0.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==2.3.6
widgetsnbextension==4.0.9
xmltodict==0.13.0
zipp==3.17.0
